Desobekify BrowserContextOptionParsing
Moves the Sobek-dependent option parsing part out of the business logic.

Once we finish, we can entirely remove the Sobek dependency.
