+++ common/browser_context_options.go
@@ -98,96 +98,6 @@ func NewBrowserContextOptions() *BrowserContextOptions {
 	}
 }
 
-// Parse parses the browser context options.
-func (b *BrowserContextOptions) Parse(ctx context.Context, opts sobek.Value) error { //nolint:cyclop,funlen,gocognit
-	rt := k6ext.Runtime(ctx)
-	if !sobekValueExists(opts) {
-		return nil
-	}
-	o := opts.ToObject(rt)
-	for _, k := range o.Keys() {
-		switch k {
-		case "acceptDownloads":
-			b.AcceptDownloads = o.Get(k).ToBoolean()
-		case "downloadsPath":
-			b.DownloadsPath = o.Get(k).String()
-		case "bypassCSP":
-			b.BypassCSP = o.Get(k).ToBoolean()
-		case "colorScheme":
-			switch ColorScheme(o.Get(k).String()) { //nolint:exhaustive
-			case "light":
-				b.ColorScheme = ColorSchemeLight
-			case "dark":
-				b.ColorScheme = ColorSchemeDark
-			default:
-				b.ColorScheme = ColorSchemeNoPreference
-			}
-		case "deviceScaleFactor":
-			b.DeviceScaleFactor = o.Get(k).ToFloat()
-		case "extraHTTPHeaders":
-			headers := o.Get(k).ToObject(rt)
-			for _, k := range headers.Keys() {
-				b.ExtraHTTPHeaders[k] = headers.Get(k).String()
-			}
-		case "geolocation":
-			geolocation := NewGeolocation()
-			if err := geolocation.Parse(ctx, o.Get(k).ToObject(rt)); err != nil {
-				return err
-			}
-			b.Geolocation = geolocation
-		case "hasTouch":
-			b.HasTouch = o.Get(k).ToBoolean()
-		case "httpCredentials":
-			credentials := NewCredentials()
-			if err := credentials.Parse(ctx, o.Get(k).ToObject(rt)); err != nil {
-				return err
-			}
-			b.HttpCredentials = credentials
-		case "ignoreHTTPSErrors":
-			b.IgnoreHTTPSErrors = o.Get(k).ToBoolean()
-		case "isMobile":
-			b.IsMobile = o.Get(k).ToBoolean()
-		case "javaScriptEnabled":
-			b.JavaScriptEnabled = o.Get(k).ToBoolean()
-		case "locale":
-			b.Locale = o.Get(k).String()
-		case "offline":
-			b.Offline = o.Get(k).ToBoolean()
-		case "permissions":
-			if ps, ok := o.Get(k).Export().([]any); ok {
-				for _, p := range ps {
-					b.Permissions = append(b.Permissions, fmt.Sprintf("%v", p))
-				}
-			}
-		case "reducedMotion":
-			switch ReducedMotion(o.Get(k).String()) { //nolint:exhaustive
-			case "reduce":
-				b.ReducedMotion = ReducedMotionReduce
-			default:
-				b.ReducedMotion = ReducedMotionNoPreference
-			}
-		case "screen":
-			screen := &Screen{}
-			if err := screen.Parse(ctx, o.Get(k).ToObject(rt)); err != nil {
-				return err
-			}
-			b.Screen = screen
-		case "timezoneID":
-			b.TimezoneID = o.Get(k).String()
-		case "userAgent":
-			b.UserAgent = o.Get(k).String()
-		case "viewport":
-			viewport := &Viewport{}
-			if err := viewport.Parse(ctx, o.Get(k).ToObject(rt)); err != nil {
-				return err
-			}
-			b.Viewport = viewport
-		}
-	}
-
-	return nil
-}
-
 // WaitForEventOptions are the options used by the browserContext.waitForEvent API.
 type WaitForEventOptions struct {
 	Timeout     time.Duration
